chore: renamed packages from digiwf* to miranum*
