+++ spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy-example/src/main/java/io/miragon/miranum/miranumdeploymentproxy/ExampleApplication.java
@@ -1,4 +1,4 @@
-package io.miragon.miranum.digiwfdeploymentproxy;
+package io.miragon.miranum.miranumdeploymentproxy;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
