+++ spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy-example/src/main/resources/application.properties
@@ -1,10 +1,10 @@
 ## Deployment Targets
 # io.miragon.digiwf.miranum-deployment-proxy.target.<target>=<baseUrl>
-io.miragon.digiwf.miranum-deployment-proxy.targets.local=http://localhost:39146
+io.miragon.miranum.digiwf-deployment-proxy.targets.local=http://localhost:39146
 
 ## Deployment Handlers
 # io.miragon.digiwf.miranum-deployment-proxy.deploymentHandlers.<target>=<handler>
-io.miragon.digiwf.miranum-deployment-proxy.deploymentHandlers.local=DryHandler
-io.miragon.digiwf.miranum-deployment-proxy.deploymentHandlers.dev=DryHandler
-io.miragon.digiwf.miranum-deployment-proxy.deploymentHandlers.test=DryHandler
-io.miragon.digiwf.miranum-deployment-proxy.deploymentHandlers.prod=DryHandler
+io.miragon.miranum.digiwf-deployment-proxy.deploymentHandlers.local=DryHandler
+io.miragon.miranum.digiwf-deployment-proxy.deploymentHandlers.dev=DryHandler
+io.miragon.miranum.digiwf-deployment-proxy.deploymentHandlers.test=DryHandler
+io.miragon.miranum.digiwf-deployment-proxy.deploymentHandlers.prod=DryHandler
