+++ spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy-example/src/test/java/io/miragon/miranum/miranumdeploymentproxy/DeploymentControllerTest.java
@@ -1,9 +1,9 @@
-package io.miragon.miranum.digiwfdeploymentproxy;
+package io.miragon.miranum.miranumdeploymentproxy;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.miragon.miranum.digiwfdeploymentproxy.dto.ArtifactDto;
-import io.miragon.miranum.digiwfdeploymentproxy.dto.DeploymentDto;
-import io.miragon.miranum.digiwfdeploymentproxy.dto.FileDto;
+import io.miragon.miranum.miranumdeploymentproxy.dto.ArtifactDto;
+import io.miragon.miranum.miranumdeploymentproxy.dto.DeploymentDto;
+import io.miragon.miranum.miranumdeploymentproxy.dto.FileDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
