+++ spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  io.miragon.miranum.digiwfdeploymentproxy.configuration.DeploymentProxyAutoConfiguration
+  io.miragon.miranum.miranumdeploymentproxy.configuration.DeploymentProxyAutoConfiguration
