--- spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy/src/main/java/io/miragon/miranum/digiwfdeploymentproxy/handler/DeploymentHandler.java
@@ -1,9 +1,0 @@
-package io.miragon.miranum.digiwfdeploymentproxy.handler;
-
-import io.miragon.miranum.digiwfdeploymentproxy.dto.DeploymentDto;
-import io.miragon.miranum.digiwfdeploymentproxy.dto.DeploymentSuccessDto;
-
-public interface DeploymentHandler {
-
-    DeploymentSuccessDto deploy(final DeploymentDto deploymentDto);
-}
