+++ spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy/src/main/java/io/miragon/miranum/miranumdeploymentproxy/dto/ArtifactDto.java
@@ -1,4 +1,4 @@
-package io.miragon.miranum.digiwfdeploymentproxy.dto;
+package io.miragon.miranum.miranumdeploymentproxy.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
