+++ spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy/src/main/java/io/miragon/miranum/miranumdeploymentproxy/dto/DeploymentDto.java
@@ -1,4 +1,4 @@
-package io.miragon.miranum.digiwfdeploymentproxy.dto;
+package io.miragon.miranum.miranumdeploymentproxy.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
@@ -6,18 +6,17 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 @Data
-@Schema(description = "File Details")
-public class FileDto {
+@Schema(description = "Deployment")
+public class DeploymentDto {
     @NotBlank
-    private String name;
-    @NotBlank
-    private String content;
-    private String extension;
-    private long size;
+    private String target;
+    @Valid
+    private ArtifactDto artifact;
 }
