+++ spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy/src/main/java/io/miragon/miranum/miranumdeploymentproxy/dto/FileDto.java
@@ -1,4 +1,4 @@
-package io.miragon.miranum.digiwfdeploymentproxy.dto;
+package io.miragon.miranum.miranumdeploymentproxy.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
@@ -6,17 +6,18 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 @Data
-@Schema(description = "Deployment")
-public class DeploymentDto {
+@Schema(description = "File Details")
+public class FileDto {
     @NotBlank
-    private String target;
-    @Valid
-    private ArtifactDto artifact;
+    private String name;
+    @NotBlank
+    private String content;
+    private String extension;
+    private long size;
 }
