+++ spring-boot-apps/miranum-deployment-proxy/miranum-deployment-proxy/src/main/java/io/miragon/miranum/miranumdeploymentproxy/handler/DeploymentHandler.java
@@ -1,0 +1,9 @@
+package io.miragon.miranum.miranumdeploymentproxy.handler;
+
+import io.miragon.miranum.miranumdeploymentproxy.dto.DeploymentDto;
+import io.miragon.miranum.miranumdeploymentproxy.dto.DeploymentSuccessDto;
+
+public interface DeploymentHandler {
+
+    DeploymentSuccessDto deploy(final DeploymentDto deploymentDto);
+}
